test: move unix_escalation tests into sibling file
